UMVE: Explicitly set Qt plugin search path
Qt5 on Windows requires a platform plugin/DLL named qwindows.dll. The default
search path seems to be hardcoded at compile-time to the location of the Qt
libraries on the machine where the code is compiled. It doesn't seem to be
possible to override the search path using qt.conf and Qt also doesn't search
the working directory of the binary. Thus we explicitly add the "qt_plugins"
folder to the search paths by calling QCoreApplication::addLibraryPath(). Then
qwindows.dll has to reside in "qt_plugins/platforms". SVG icons then work as
well if qsvg.dll is available in "qt_plugins/imageformats".

Related bug report in Qt's official bug tracker:
https://bugreports.qt-project.org/browse/QTBUG-28766