SFM: Fix compilation on Visual Studio in debug mode
In debug mode, Visual Studio checks at run-time whether there is enough
space in the container of the output iterator, if possible. Visual Studio's
std::copy() interprets cam.m[0] as a float[3] but the number of elements to be
copied is 9 (pose.R) which causes the run-time check to fail. This is fixed by
explicitly passing a pointer as the output iterator to omit the range check.