OGL: Explicitly set primitive type in MeshRenderer::set_mesh()
There is a potential bug in the MeshRenderer class when calling
MeshRenderer::set_mesh() with a point cloud. The primitive type is left at
GL_TRIANGLES, as initialized in the ctor of VertexArray of which
MeshRenderer is derived from. Then, when MeshRenderer/VertexArray::draw() is
called, triangles are drawn instead of points. Explicitly setting the
primitive type in MeshRenderer::set_mesh() fixes the issue.