Add Libexecinfo flag for alpine + create Dockerfile
Alpine doesn't include the backtrace functionality of glibc, which is
used in libs/util/system.cc. Libexecinfo provides the missing function
symbols for Alpine and FreeBSD distros. Adding a check in Makefile.inc
for /etc/alpine-release will determine if the library needs to be
linked.

Provide a Dockerfile to validate the build on Alpine. The Dockerfile
provides a minimal set of dependencies needed for compilation.