Check for GLIBC rather than GNUC for compatibility
The current compatibility checks in the libs/util/system.cc
check for the __GNUC__ flag when deciding whether to include
the execinfo.h header and backtrace functionalityt. The
__GNUC__ flag returns true for distros that do not inlude glibc
which will cause compilation issues for musl libc.

The check now looks for __GLIBC__ and will properly exclude the
functionality from distros using a different libc implementation.